Added generic sonar scanner jenkins file
This is generic template for running sonar scanner. The intended usages is to
create one pipeline to take that jenkins file path and then per each source
tree in Jenkinsfile.BUILD trigger the sonar build pipeline at the end.

By doing do we save a lot of checking repository time and effort.

This has some assumption about the Minicover folders that projects need to
have. See the source for details.