Back to ansible 2.8.1
The error made us to roll back seems not releated to ansible as I test ansible
2.8.1 and it honours the `library` path in role thus it should use the same
win_iis_web_application modules no matter which version of anisble is.

it might be that it always gives error in some special condition and might be
that before powershell module tries to update it gets the current web app
attributes and verify it - and the folder path is already clean up earlier.

We might be able to fix it by looking to the way we clean up old deployment
folder instead.