Remove the procedure at agent
Looks like it is not easy with jenkins. We have to run the harvest in the
master only and need to make the downstream job being triggered by the main
build job to just collect the log file. Maybe wget, etc not sure how to archive
it yet.

The CompressLog() property seems not to work.