Added get_build_properties
Similar to the other function but this one filter not only by parameters but also the job descriptions.

Use case is in act2, when we use the one of the job param composed by other
variables such as $BUILD_NUMBER/$BUILD_NAME - when querying via that param we
do not get the final string, but the '$BUILD_NUMBER/$BUILD_NAME' in origin.

For these cases the build description might be a source of truth to query for.